Refactor GutterAskWrapper into its own file
